Day 1: Add solutions
I solved day 1 in C++, but I've added a C# version here after the fact.

Difference: Part 2 was originally solved in a "smarter", but more
complex and error-prone way. The solution split "full rotations" (=100
turns) from the remaining "partial rotations" (<100 turns). Every full
rotation is a click, and a partial rotation is a click if we end up at
a position <0 or >=100. But then it's harder to deal with edge cases
where the dial position started on zero or ends up at zero.

I watched Jonathan Paulson on YouTube simply simulate every turn
movement one at a time - surely really inefficient, but it really does
not matter here and makes for a much easier-to-implement solution.