Day 12: Add solution
In the morning of day 12, I only implemented the area comparison check,
but never tried the answer on the website. My solution didn't work for
the example input, so I thought, clearly you would need to implement at
least _some_ algorithm for packing of irregular shapes, even if it
wasn't perfect.

...Nope. A work colleague gave me a hint. The "naive" solution is
entirely sufficient to solve the puzzle. Go figure.

All I've done since then is to clean up the code a bit and add a
"tolerance" to the area check, so that the solution also solves
correctly for the example input.